refactor: 데이터에 맞는 파일명으로 변경(#177)
