+++ src/components/recruitment-detail/recuitment-lecture/LectureCard.tsx
