+++ src/components/recruitment-detail/recuitment-lecture/RecuitmentLecture.tsx
